zsh: profile: align environ invocation with .profile
Side note: Damn I'm stupid. The [ "$PREFIX" ] avoided the previously
described problem nicely even when this was in zshenv, as environ always
sets $PREFIX.

But again, let's keep it aligned with bash's behavior for the time
being. I might use zshenv, $BASH_ENV, and $ENV to re-add this kind of
"insurance" later...