xprofile: just remove globalenv outright
Apparently I don't actually use it, and it's somewhat broken in cron
anyway. (Not sure if it is, really. But if I do `keyctl new_session`, it
can find the keys but not read them due to permission errors. I wonder
if cron calls pam_keyinit? If it does, then it'd get a link to the user
keyring and globalenv would work fine... Still.)