xinitrc: load profile in two stages
Load the general 'profile' early, because it does not depend on any
daemons and starts nothing persistent. Rather, the daemons depend on
environment variables that ~/.profile might set -- in particular $LANG
for dbus-daemon, otherwise gnome-terminal-server will refuse to start.

Keep the loading of 'xprofile' late, though, because it does depend on
those daemons (e.g. it might start notification-daemon).