bash/prompt: change foo.bar to foo:bar in some places
There are two uses for subformats:

 - different *parts* of an item, e.g. prefix/suffix
 - different *modes* of an item, e.g. self/root

For different parts, I usually have fmt[base] defined so it makes sense
for fmt[base.sub] to fall back to that.

But for different modes, it's the opposite – fmt[base] is an alias to
one mode or another – so if the mode is missing the format, fallback
actually causes an infinite loop.

So use different separator that avoids the fallback.