bashrc: always unexport LINES and COLUMNS
It turns out bash's creation of $LINES is non-deterministic – it doesn't
use ioctl(TIOCGWINSZ), it sends the "report window size" sequence and
doesn't actually wait for the terminal to respond.

If the terminal is a bit slow and happens to respond when bash is
already in the middle of environ's 'allexport' section, then $LINES will
be exported as well, which is what causes the ncurses resizing problem.