remove mailrc
Using heirloom-mailx for IMAP was kind of okay, but its successor s-nail
has turned into a monster and I haven't actually used it for years. (And
who knows if they won't decide to completely remove IMAP support
*again*?)

Instead of trying to update the config for the current version
(s/sendmail/mta/, signature → on-compose-splice) just remove it
entirely.