tmux: remove ineffective preview toggle
It only matches 2.* so it stopped working when tmux 3.0 came out and I
never got around to fixing it. In the end, the previews are actually
kinda useful.

(If I ever want to disable previews again, I can just do a normal bind
as compat with <= 2.5 is no longer a concern -- even buster has 2.8,
with 3.1 in backports.)