bash: "fix" etsn
As PS remoting only reconfigures the current process instead of starting
a new one (as would be the case with e.g. SSH), the next 'exit' actually
exits from the remoting session...

(That is, Enter-PSSession switches the interpreter to a new runspace and
all commands that follow are processed in that runspace. That's also the
reason the -NoExit is required, as without it, PowerShell will just hang
waiting for the remote runspace to exit.)